Purge common twi (#163)
* init update

* spi, twi: поддержка работы нескольких портов

* build fix

* remove config: arm_allwt507_ctlstyle_helperboard_t507_ra4asn.h

* Build fix. LVGL update

* build fix

+++ src/gui/gui_user.c
@@ -1241,7 +1241,7 @@ void window_memory_process(void)
 		{
 			local_snprintf_P(name, NAME_ARRAY_SIZE, "btn_memory_%d", i);
 
-			uint32_t freq = load_mems(i, 0);
+			long freq = load_mems(i, 0);
 			if (freq > 0)
 				local_snprintf_P(text, NAME_ARRAY_SIZE, "%ldk", freq / 1000);
 			else
